Fix login method in controller and jsp files
