+++ src/main/java/app/web/view/UserViewController.java
@@ -6,37 +6,53 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
 
 @Controller
 @RequestMapping("/users")
 @RequiredArgsConstructor
 public class UserViewController {
 
-    private final String ATTR_LANGUAGE ="lang";
-    private final String ATTR_USER ="user";
-    private final String ATTR_MESSAGE ="user";
-    private final String MESSAGE_LOGIN_REQUEST ="Please login";
+    private final String ATTR_LANGUAGE = "lang";
+    private final String ATTR_USER = "user";
+    private final String ATTR_MESSAGE = "user";
+    private final String MESSAGE_LOGIN_REQUEST = "Please login";
+    private final String MESSAGE_SUCCEFULLY_LOGIN = "Succesfully login";
 
-    private final String INDEX_PAGE="index";
-    private final String LOGIN_FORM="user/loginForm";
-    private final String USER_PAGE="user/user";
-    private final String USERS_PAGE="user/users";
-    private final String REGISTER_PAGE="user/register";
-    private final String PROFILE_PAGE="user/profile";
-    private final String EDIT_PAGE="user/edit";
+    private final String INDEX_PAGE = "index";
+    private final String LOGIN_FORM = "user/loginForm";
+    private final String USER_PAGE = "user/user";
+    private final String USERS_PAGE = "user/users";
+    private final String REGISTER_PAGE = "user/register";
+    private final String PROFILE_PAGE = "user/profile";
+    private final String EDIT_PAGE = "user/edit";
 
     private final UserService userService;
+
+    @ModelAttribute("user")
+    public UserDto userDto() {
+        return new UserDto();
+    }
 
     @GetMapping("/login")
     public String loginForm() {
         return LOGIN_FORM;
     }
 
+    @PostMapping("/login")
+    public String login(@ModelAttribute @Valid UserDto userDto, HttpSession session, Model model) {
+        UserDto user = userService.login(userDto.getLogin(), userDto.getPassword());
+        session.setAttribute(ATTR_USER, user);
+        model.addAttribute(ATTR_MESSAGE, MESSAGE_SUCCEFULLY_LOGIN);
+        return INDEX_PAGE;
+    }
 
     @GetMapping("/logout")
     public String logout(HttpServletRequest req) {
